fix(payments): separate modal and carousel
